refactor(product, utils): changed name and type of  constructor parameter.
Renamed parameter (qtdPerPackage -> quantityPerPackage) and change its type (double -> int). The super class SampleContainer will cast the int for double to ensure coherence and block illogical values for all SampleContainer subclasses

+++ apps/web-api-java/src/main/java/br/com/biosecure/model/product/SampleContainer.java
@@ -9,9 +9,9 @@ public abstract class SampleContainer extends  Product {
     private final Material material;
     private final double capacityMiliLiters;
 
-    public SampleContainer(String name, double price, String manufacturer, String batchNumber, LocalDate expirationDate, PackagingType packagingType, double quantityPerPackage, SterilizationMethod sterilizationMethod, ClosingMethod closingMethod, Material materialType, double capacityMiliLiters) {
+    public SampleContainer(String name, double price, String manufacturer, String batchNumber, LocalDate expirationDate, PackagingType packagingType, int quantityPerPackage, SterilizationMethod sterilizationMethod, ClosingMethod closingMethod, Material materialType, double capacityMiliLiters) {
         
-        super(name, price, manufacturer, batchNumber, expirationDate, packagingType, MeasureUnit.UN, quantityPerPackage);
+        super(name, price, manufacturer, batchNumber, expirationDate, packagingType, MeasureUnit.UN, (double) quantityPerPackage);
 
         if (capacityMiliLiters <= 0 || capacityMiliLiters > 99999) {
             throw new InvalidProductAttributeException("capacity");
